requirements.txt
requests>=2.31.0
dnspython>=2.6.1
colorama>=0.4.6
